Abgabe_VDA/Django-Server/VDA/requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
Click==7.0
cryptography==2.3.1
Django==2.1.4
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
idna==2.7
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
linecache2==1.0.0
MarkupSafe==1.0
mysqlclient==1.3.14
mysqlclient-deps==0.0.2
mysql-connector-python==8.0.13
Pillow==5.3.0
pip==18.1
protobuf==3.6.1
pycparser==2.18
PyMySQL==0.9.2
PyMysqlDB==0.0.2
pytz==2018.5
pywebuml==0.3.0
RPi.GPIO-def==0.2
setuptools==39.0.1
six==1.11.0
SQLAlchemy==1.2.13
traceback2==1.4.0
unittest2==1.1.0
Werkzeug==0.14.1
